refactor: 스키마 분할
