+++ package.json
@@ -20,9 +20,13 @@
     },
     "homepage": "https://github.com/EumMericano-Study/instaclone-backend#readme",
     "dependencies": {
+        "@graphql-tools/load-files": "^6.5.3",
+        "@graphql-tools/merge": "^8.2.1",
+        "@graphql-tools/schema": "^8.3.1",
         "@prisma/client": "^3.8.1",
         "apollo-server": "^3.6.1",
         "graphql": "^16.2.0",
+        "graphql-tools": "^8.2.0",
         "typescript": "^4.5.4"
     },
     "devDependencies": {
